requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

torch>=2.1.0
torchvision>=0.16.0

transformers>=4.46.0
datasets>=3.0.0
safetensors>=0.4.2

ultralytics>=8.2.0

matplotlib>=3.7.0
seaborn>=0.13.0

Pillow>=10.0.0
opencv-python>=4.8.0
openpyxl>=3.1.0

PyYAML>=6.0.0
tqdm>=4.66.0

accelerate>=0.30.0
timm>=1.0.0
evaluate>=0.4.1
